capstone/pyproject.toml
[tool.poetry]
name = "capstone"
version = "0.1.0"
description = ""
authors = ["Tzvi Friedman <tadwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
gunicorn = "^21.2.0"
flask = "^3.0.0"
pillow = "^10.1.0"
requests = "^2.31.0"
scikit-learn = "^1.3.2"
apig-wsgi = "^2.18.0"

[tool.poetry.group.dev.dependencies]
tensorflow-macos = { version = "^2.15.0", markers = "platform_system=='Darwin' and platform_machine=='arm64'" }
tensorflow = "^2.15.0"
ipykernel = "^6.27.0"
matplotlib = "^3.8.0"
pandas = "^2.1.3"
numpy = "^1.26.1"
seaborn = "^0.13.1"
pytest = "^7.4.3"
aws-cdk-lib = "^2.110.1"
constructs = "^10.3.0"

[tool.poetry.group.lambda.dependencies]
tflite_runtime = { url = "https://github.com/alexeygrigorev/tflite-aws-lambda/blob/main/tflite/tflite_runtime-2.14.0-cp311-cp311-linux_x86_64.whl/?raw=true"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
